Edit Functionality Working
Edit Functionality now finally fulling working.

Can new display genres that were already associated with a book for ease of use.

can handle cases where a genre is added and removed.

Changes to SQLAlchemy models
GenreBooks is now a Table not a class which makes it more simple  from a schema perspective and a query prespective