Edit Book Refactor + View all books
Edit book template now has the genres aligned with all other elements and now displays the checkboxes in rows and columns so its uses the whole rather then creating a long list of single checkboxes. Added a Cancel button to back out of making changes. This redirects user to the last page they were on.

Added a View Books route and template which shows you ALL the current books in the system, includes edit button tot take to edit page. Added navigation to this page in Base.html template.